requirements.txt
google-generativeai
openai
langchain
pymongo
redis
fastapi
uvicorn
tiktoken
python-dotenv